+++ src/test/java/org/apache/tiles/request/jakarta/servlet/ExternalWriterHttpServletResponseTest.java
@@ -26,7 +26,6 @@ import java.io.Writer;
 
 import jakarta.servlet.http.HttpServletResponse;
 
-import org.apache.tiles.request.jakarta.servlet.ExternalWriterHttpServletResponse;
 import org.junit.Test;
 
 /**