+++ src/test/java/org/apache/tiles/request/jakarta/servlet/extractor/InitParameterExtractorTest.java
@@ -22,7 +22,6 @@ import java.util.Enumeration;
 
 import jakarta.servlet.ServletContext;
 
-import org.apache.tiles.request.jakarta.servlet.extractor.InitParameterExtractor;
 import org.junit.Before;
 import org.junit.Test;
 
@@ -56,7 +55,6 @@ public class InitParameterExtractorTest {
     /**
      * Test method for {@link org.apache.tiles.request.servlet.extractor.InitParameterExtractor#getKeys()}.
      */
-    @SuppressWarnings("unchecked")
     @Test
     public void testGetKeys() {
         Enumeration<String> keys = createMock(Enumeration.class);